% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baum.R
\name{baum}
\alias{baum}
\title{Implements the Baum-Welch algorithm to guess initial distribution, transition probabilities, and
second row of the emission matrix}
\usage{
baum(em_row, y, T_end, K_end)
}
\arguments{
\item{em_row}{the row for the fair state}

\item{y}{the history of observable actions}

\item{T_end}{number of timesteps in y}

\item{K_end}{number of iterations of times the Baum-Welch algorithm is looped}
}
\description{
Implements the Baum-Welch algorithm to guess initial distribution, transition probabilities, and
second row of the emission matrix
}
