% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_custom.R
\name{EM_custom}
\alias{EM_custom}
\title{Expectation-Maximization function for the blood type example}
\usage{
EM_custom(n, p_vec, T_end)
}
\arguments{
\item{n}{is the phenotype vector (n_a, n_b, n_ab, n_o)}

\item{p_vec}{is the allele vector (p_a, p_b, p_o)}

\item{T_end}{indicates how often the for-loop iterates}
}
\description{
Expectation-Maximization function for the blood type example
}
