% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenStateInference.R
\name{hiddenStateInference}
\alias{hiddenStateInference}
\title{Marginal probability calculator for hidden states, given forward and backward HMM solver outputs
outputs an N_s x N_T matrix, where N_s is the number of hidden states and
N_T is the number of timesteps}
\usage{
hiddenStateInference(a, b, T_end)
}
\arguments{
\item{a}{forward HMM solver output}

\item{b}{backward HMM solver output}

\item{T_end}{total number of timesteps}
}
\description{
Marginal probability calculator for hidden states, given forward and backward HMM solver outputs
outputs an N_s x N_T matrix, where N_s is the number of hidden states and
N_T is the number of timesteps
}
