% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMbackward.R
\name{HMMbackward}
\alias{HMMbackward}
\title{Backward algorithm for HMM solving
outputs an N_s x N_T matrix, where N_s is the number of hidden states and
N_T is the number of timesteps}
\usage{
HMMbackward(y, P, E, v0)
}
\arguments{
\item{y}{observable actions taken}

\item{P}{transition probabilities for hidden states}

\item{E}{emission matrix for observable actions, given hidden actions}

\item{v0}{initial guess}
}
\description{
Backward algorithm for HMM solving
outputs an N_s x N_T matrix, where N_s is the number of hidden states and
N_T is the number of timesteps
}
