% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMforward.R
\name{HMMforward}
\alias{HMMforward}
\title{Forward algorithm for HMM solving
outputs an N_s x N_T matrix, where N_s is the number of hidden states and
N_T is the number of timesteps}
\usage{
HMMforward(y, P, E, v0)
}
\arguments{
\item{y}{observable actions taken}

\item{P}{transition probabilities for hidden states}

\item{E}{emission matrix for observable actions, given hidden actions}

\item{v0}{initial guess}
}
\description{
Forward algorithm for HMM solving
outputs an N_s x N_T matrix, where N_s is the number of hidden states and
N_T is the number of timesteps
}
