% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logres_FS.R
\name{logres_FS}
\alias{logres_FS}
\title{Logistic regression optimization with Fisher-scoring}
\usage{
logres_FS(x, y, theta, alpha, T_end)
}
\arguments{
\item{x}{the covariate variables}

\item{y}{the response variable}

\item{theta}{the initial parameter estimate}

\item{alpha}{the learning rate}

\item{T_end}{the number of iterations for which the optimization is allowed to run}
}
\description{
Logistic regression optimization with Fisher-scoring
}
