% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logres_GD.R
\name{logres_GD}
\alias{logres_GD}
\title{Logistic regression optimization with gradient-descent}
\usage{
logres_GD(x, y, theta, alpha, T_end)
}
\arguments{
\item{x}{the covariate variables}

\item{y}{the response variable}

\item{theta}{the initial parameter estimate}

\item{alpha}{the learning rate}

\item{T_end}{the number of iterations for which the optimization is allowed to run}
}
\description{
Logistic regression optimization with gradient-descent
}
