% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult2mv.R
\name{mult2mv}
\alias{mult2mv}
\title{Compare two different linear mappings.}
\usage{
mult2mv(A, B, x, swtch)
}
\arguments{
\item{A}{The left matrix. (m x p)}

\item{B}{The right matrix. (p x n)}

\item{x}{The vector. (n x 1)}

\item{swtch}{Boolean Flag to use one or the other associative ordering of multiplication.}
}
\description{
Multiplying matrices and vectors is associative. This function performs multiplies matrix A (mxp), matrix B(pxn), and vector x (nx1) in two different ways.
}
\examples{
mult2mv(matrix(1:4,2,2),matrix(5:8,2,2),c(2,1),TRUE)
mult2mv(matrix(1:8,2,4),matrix(8:19,4,3),c(3,2,1),FALSE)
}
