% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNsampler.R
\name{MVNsampler}
\alias{MVNsampler}
\title{Return N samples from the a multivariate normal (MVN) density function}
\usage{
MVNsampler(N, m, S)
}
\arguments{
\item{N}{the number of realizations (samples)}

\item{m}{mu, the mean of the multivariate normal (MVN) density function}

\item{S}{sigma, the covariance matrix of the MVN. must be positive definite}
}
\description{
Return N samples from the a multivariate normal (MVN) density function
}
\examples{
mu = runif(2)
cv.pre = matrix(runif(4,min=0.01,max=1.01),nrow=2)
cv = 1/2 * (cv.pre + cv.pre)
samples = MVNsampler(mu,cv,)
}
