% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_student.R
\name{QR_student}
\alias{QR_student}
\title{QR decomposition-based OLS}
\usage{
QR_student(x, y)
}
\arguments{
\item{y}{response}

\item{X}{covariates}
}
\value{
regression coefficients c
}
\description{
QR decomposition-based OLS
}
