% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD_student.R
\name{SVD_student}
\alias{SVD_student}
\title{SVD-decomposition based OLS}
\usage{
SVD_student(x, y)
}
\arguments{
\item{y}{response}

\item{X}{covariates}
}
\value{
regression coefficients c
}
\description{
SVD-decomposition based OLS
}
